+++ miru/graphql/queries.py
@@ -36,19 +36,19 @@ class AnimeEntryListResults(graphene.ObjectType):
 
 class Query(graphene.ObjectType):
 
-    anime_by_id = graphene.Field(AnimeType, id=graphene.ID(required=True))
-    characters_by_anime = graphene.List(AnimeCharacterType, id=graphene.ID(required=True))
+    anime_by_id = graphene.Field(AnimeType, anime_id=graphene.ID(required=True))
+    characters_by_anime = graphene.List(AnimeCharacterType, anime_id=graphene.ID(required=True))
     anime_by_category = graphene.List(AnimeType, category=graphene.String(required=True), count=graphene.Int(required=False))
     search_anime = graphene.Field(AnimeFilterResults, filters=AnimeFilterInput(), sort=MediaSortInput(), pagination=PaginationInput())
     get_anime_list = graphene.Field(AnimeEntryListResults, user_id=graphene.ID(required=True))
     get_anime_list_entry = graphene.Field(AnimeListEntryType, user_id=graphene.ID(required=True), anime_id=graphene.ID(required=True))
     get_anime_episodes = graphene.List(AnimeEpisodeType, anime_id=graphene.ID(required=True))
 
-    def resolve_anime_by_id(self, _info, id):
-        return MiruService.get_anime_by_id(id)
+    def resolve_anime_by_id(self, _info, anime_id):
+        return MiruService.get_anime_by_id(anime_id)
     
-    def resolve_characters_by_anime(self, _info, id):
-        return MiruService.get_characters_by_anime(id)
+    def resolve_characters_by_anime(self, _info, anime_id):
+        return MiruService.get_characters_by_anime(anime_id)
     
     def resolve_anime_by_category(self, _info, category, count):
         return MiruService.get_anime_by_category(f'-{category}', count)