+++ talent/graphql/query.py
@@ -1,13 +1,14 @@
 import graphene
 
-from .schema import (
-    CharacterType,
-    VoiceActorType
-)
 from asobu.graphql.schema import GameCharacterType
 from miru.graphql.schema import AnimeCharacterType
 from talent.service.character import CharacterService
 from talent.service.voice_actor import VoiceActorService
+
+from .schema import (
+    CharacterType,
+    VoiceActorType
+)
 
 class VoiceActorResults(graphene.ObjectType):
     voice_actor = graphene.Field(VoiceActorType)
@@ -16,14 +17,14 @@ class VoiceActorResults(graphene.ObjectType):
 
 class Query(graphene.ObjectType):
 
-    character_by_id = graphene.Field(CharacterType, id=graphene.Int(required=True))
-    voice_actor_by_id = graphene.Field(VoiceActorResults, id=graphene.ID(required=True))
+    character_by_id = graphene.Field(CharacterType, character_id=graphene.Int(required=True))
+    voice_actor_by_id = graphene.Field(VoiceActorResults, va_id=graphene.ID(required=True))
 
-    def resolve_character_by_id(self, info, id):
-        return CharacterService.get_character_by_id(id)
+    def resolve_character_by_id(self, _info, character_id):
+        return CharacterService.get_character_by_id(character_id)
     
-    def resolve_voice_actor_by_id(self, info, id):
-        voice_actor, characters =  VoiceActorService.get_voice_actor_by_id(id, True)
+    def resolve_voice_actor_by_id(self, _info, va_id):
+        voice_actor, characters =  VoiceActorService.get_voice_actor_by_id(va_id, True)
 
         related_anime = characters['animes']
         related_games = characters['games']