+++ talent/repository/character.py
@@ -5,8 +5,8 @@ from talent.models import (
 class CharacterRepository:
 
     @staticmethod
-    def get_character_by_id(id):
+    def get_character_by_id(character_id):
         try:
-            return Character.objects.get(id=id)
+            return Character.objects.get(id=character_id)
         except Character.DoesNotExist:
             return None